requirements.txt
asgiref==3.7.2
Django==5.0.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
python-dotenv==1.0.1
pytz==2024.1
ruff==0.3.0
sqlparse==0.4.4
